Detalles validaciones

+++ src/main/java/com/grupo3/oficio/service/users/ClienteService.java
@@ -2,6 +2,7 @@ package com.grupo3.oficio.service.users;
 
 import com.grupo3.oficio.model.users.Cliente;
 import com.grupo3.oficio.repository.users.ClienteRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -10,7 +11,7 @@ import java.util.NoSuchElementException;
 
 @Service
 public class ClienteService {
-    private ClienteRepository clienteRepository;
+    private final ClienteRepository clienteRepository;
 
     public ClienteService(ClienteRepository clienteRepository) { this.clienteRepository = clienteRepository; }
 
@@ -87,6 +88,9 @@ public class ClienteService {
         if (cliente.getDni() == null || !cliente.getDni().matches("\\d{7,8}")) {
             throw new IllegalArgumentException("DNI inválido");
         }
+        if (!clienteRepository.existsById(id)) {
+            throw new EntityNotFoundException("El cliente con ID " + id + " no existe");
+        }
 
         cliente.setId(id);
         clienteRepository.save(cliente);
@@ -96,10 +100,11 @@ public class ClienteService {
 
         //delete
     public void borrar(Integer id) {
-        clienteRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("ERROR ClienteService/borrar, NO " + id));
         if (id == null) {
             throw new NullPointerException("ERROR ClienteService/borrar, ID nulo");
+        }
+        if (!clienteRepository.existsById(id)) {
+            throw new EntityNotFoundException("El cliente con ID " + id + " no existe");
         }
 
         clienteRepository.deleteById(id);
